Move installation instructions in to distro-specific folders.
For previous ROS 2 releases we've stacked installation instructions on
top of each other and used prose to indicate when instructions pertain
to or exclude certain releases.

As we add more releases the prose gets more convoluted and harder for
readers to follow. Instead we'll create a bit of duplication in order to
keep the installation instructions for Crystal (and earlier) intact
while making edits for Dashing in a separate parallel directory
(forthcoming).

I have addressed sphinx build warnings locally and have changed several
instances of `../Tutorials/...`  to `/Tutorials/...` so that the
relative depth of tutorials and the installation docs isn't significant.
At least testing locally with `make html` these are rendered into html
as relative links so they work regardless of the directory structure
above the sphinx root.